requirements.txt
tensorflow-gpu==1.3
keras==2.0.8
imgaug
opencv-python
h5py
pyqt==5.6
